% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_internet.R
\docType{data}
\name{data_internet}
\alias{data_internet}
\title{data_internet}
\format{
An object of class \code{data.frame} with 6 rows and 17 columns.
}
\usage{
data_internet
}
\description{
data_internet
}
\keyword{datasets}
